requirements.txt
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.0
pytest-cov==5.0.0
ruff==0.4.4
